Extracted GuidesInspector from components.js and switched to jsx
